azurefunctions/nasdaq/requirements.txt
azure-functions
azure-storage==0.36.0